cost optimize for small data
