+++ src/data-pipeline/lambda/emr-job-submitter/emr-client-util.ts
@@ -506,11 +506,22 @@ export function getEstimatedSparkConfig(objectsInfo: ObjectsInfo): CustomSparkCo
   let inputRePartitions = 10;
 
   if (objectsInfo.sizeTotal < 1 * size_1G) {
-    logger.info('use default settings');
+    driverCore = 2;
+    driverMem = 7;
+    executorCore = 4;
+    executorMem = 14;
+    executorDisk = 20;
+    initialExecutors = 2;
+  } else if (objectsInfo.sizeTotal < 5 * size_1G) {
+    executorCore = 4;
+    executorMem = 14;
+    executorDisk = 20;
+    initialExecutors = 3;
   } else if (objectsInfo.sizeTotal < 10 * size_1G) {
     executorCore = 8;
     executorMem = 50;
     executorDisk = 50;
+    initialExecutors = 3;
   } else if (objectsInfo.sizeTotal < 50 * size_1G) {
     driverCore = 8;
     driverMem = 50;
